+++ core.lua
@@ -1,35 +1,17 @@
-local addonName = "BuffDuty"
-BuffDuty = LibStub("AceAddon-3.0"):NewAddon(addonName, "AceConsole-3.0")
-local command = "buffduty"
-
-local SAY_CHANNEL_TYPE = "SAY"
-local RAID_CHANNEL_TYPE = "RAID"
-local BG_CHANNEL_TYPE = "INSTANCE_CHAT"
+BuffDuty = LibStub("AceAddon-3.0"):NewAddon("BuffDuty", "AceConsole-3.0")
+local CHAT_COMMAND = "buffduty"
 
 function BuffDuty:OnInitialize()
-    self:RegisterChatCommand(command, "Command")
+    self:RegisterChatCommand(CHAT_COMMAND, "Command")
 end
 
-local function selectChannelType(input)
-    if (input == "s" or input == "say") then
-       return SAY_CHANNEL_TYPE
-    end
-
-    -- if (input == "r" or input == "raid") then
-
-    local inInstance, instanceType = IsInInstance()
-    -- if a person is in BG - use the battleground channel
-    if(inInstance and instanceType == "pvp") then
-        return BG_CHANNEL_TYPE
-    end
-
-    return RAID_CHANNEL_TYPE
-end
 
 local function executeLogic(input)
-    local channel_type = selectChannelType(input)
-    local duties = BuffDuty:getDutiesTable()
-    BuffDuty:printDuties(duties, channel_type)
+    local class, ch_type, channel_name = LibStub("AceConsole-3.0"):GetArgs(input, 3)
+    class, ch_type, channel_name = BuffDuty:convertArgs(class, ch_type, channel_name)
+    BuffDuty:validateArgs(class, ch_type, channel_name)
+    local duties = BuffDuty:getDutiesTable(class)
+    BuffDuty:printDuties(class, duties, ch_type, channel_name)
 end
 
 function BuffDuty:Command(input)